apps/web-api/app/emergency-backend/requirements.txt
# --- Core Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.6

# --- Database (MongoDB) ---
motor==3.6.0
pymongo==4.9.1

# --- Data Validation & Env ---
pydantic==2.9.2
python-dotenv==1.0.1
orjson==3.10.7

# --- Authentication & Security ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# --- Scheduling / Mock Feed Generator ---
apscheduler==3.10.4

# --- Utils ---
requests==2.32.3
typing_extensions==4.12.2

# --- Optional (Routing/Tests) ---
googlemaps==4.10.0
pytest==8.3.3
httpx==0.27.2
